mon: Add a new MonitorObjectStore to replace the annoying MonitorStore
We already have an ObjectStore interface which is awesome, and yet we
still rely on the MonitorStore class to do all things related with storing
values for the Monitor-related classes. And the MonitorStore does some
stuff that should be prohibited by law.

Creating a MonitorObjectStore is the first step to replace the existing
MonitorStore, which not only brings joy to the world, but also allows us
to have only one kind of store instead of two.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>