mon: fix slurp latest race
It is possible for the latest version to get out in front of the
last_committed version:

 a- start slurping
 a- slurp a bunch of states, through X
 a- get them back, write them out
 b- monitor commits many new states
 a- slurp latest, X+100 say, but only get some of those states due to the
    slurp per-message byte limit
 a- write latest + some (but not all) prior states
 a- call back into slurp(), update_from_paxos(), trigger assert

This fix ensures that we make note of the source's new latest, so that on
the next pass through slurp() we will grab any missing states.

We *also* explicitly require that we get everything up through what we have
stashed, in defense against some future kludging that might only require we
nearly (but not completely) in sync before finishing the slurp.

Fixes: #2379
Signed-off-by: Sage Weil <sage@inktank.com>
Reviewed-by: Greg Farnum <greg@inktank.com>