mon: Paxos: Get rid of the is_leader() function and add an is_lease_valid() function
The Paxos::is_leader() wasn't being used anywhere. Get rid of it for the
sake of simplicity and because it if far more clear if we call the
Monitor::is_leader() function instead.

The Paxos::is_lease_valid() will return true if the lease is valid. It is
useful to consider that on the PaxosService::is_{readable,writeable}
functions.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>